+++ modules/codegen/src/smithy4s/codegen/internals/SmithyToIR.scala
@@ -157,7 +157,7 @@ private[codegen] class SmithyToIR(
 
         shape.tpe.flatMap {
           case Type.Alias(_, name, tpe: Type.ExternalType, isUnwrapped) =>
-            val newHints = hints.filterNot(_ == tpe.refinementHint)
+            val newHints = hints.filterNot(_ sameNativeTrait tpe.refinementHint)
             TypeAlias(
               shape.getId(),
               name,
@@ -686,8 +686,9 @@ private[codegen] class SmithyToIR(
       def getHints(tpe: Type, shape: Shape): List[Hint] = {
         val h = hints(shape)
         tpe match {
-          case e: Type.ExternalType => h.filterNot(_ == e.refinementHint)
-          case _                    => h
+          case e: Type.ExternalType =>
+            h.filterNot(_ sameNativeTrait e.refinementHint)
+          case _ => h
         }
       }
 
@@ -701,13 +702,7 @@ private[codegen] class SmithyToIR(
           }
           .map { tpe =>
             val _hints = hints(x)
-            val memberHints = {
-              val h = hints(x.getMember())
-              tpe match {
-                case e: Type.ExternalType => h.filterNot(_ == e.refinementHint)
-                case _                    => h
-              }
-            }
+            val memberHints = getHints(tpe, x.getMember)
             if (_hints.contains(Hint.UniqueItems)) {
               Type.Collection(CollectionType.Set, tpe, memberHints)
             } else if (_hints.contains(Hint.SpecializedList.Vector)) {
@@ -1074,7 +1069,7 @@ private[codegen] class SmithyToIR(
 
     def tpe: Option[Type] = shape.accept(toType)
 
-    private def fieldsInternal(hintsExtractor: Shape => List[Hint]) = {
+    def fields: List[Field] = {
       val noDefault =
         if (defaultRenderMode == DefaultRenderMode.NoDefaults)
           List(Hint.NoDefault)
@@ -1093,13 +1088,13 @@ private[codegen] class SmithyToIR(
             member.getMemberName(),
             member.tpe,
             modifier,
-            hintsExtractor(member) ++ default ++ noDefault
+            hints(member) ++ default ++ noDefault
           )
         }
         .zipWithIndex
         .collect {
           case ((name, Some(tpe: Type.ExternalType), modifier, hints), index) =>
-            val newHints = hints.filterNot(_ == tpe.refinementHint)
+            val newHints = hints.filterNot(_ sameNativeTrait tpe.refinementHint)
             Field(name, tpe, modifier, index, newHints)
           case ((name, Some(tpe), modifier, hints), index) =>
             Field(name, tpe, modifier, index, hints)
@@ -1116,20 +1111,6 @@ private[codegen] class SmithyToIR(
         case DefaultRenderMode.NoDefaults => result
       }
     }
-
-    /**
-      * Should be used when calculating schema for a structure.
-      *
-      * See https://github.com/disneystreaming/smithy4s/issues/1296 for details.
-      */
-    def fields: List[Field] = fieldsInternal(hintsExtractor = hints)
-
-    /**
-      * Should be used only on the call site
-      * of the trait application where there is no need to call `unfoldTrait` for every hint of the trait.
-      */
-    def getFieldsPlain: List[Field] =
-      fieldsInternal(hintsExtractor = _ => List.empty)
 
     def alts = {
       shape
@@ -1154,7 +1135,7 @@ private[codegen] class SmithyToIR(
             Alt(
               name,
               UnionMember.TypeCase(tpe),
-              h.filterNot(_ == tpe.refinementHint)
+              h.filterNot(_ sameNativeTrait tpe.refinementHint)
             )
           case (name, Some(Right(tpe)), h) =>
             Alt(name, UnionMember.TypeCase(tpe), h)
@@ -1280,7 +1261,10 @@ private[codegen] class SmithyToIR(
   }
 
   private def unfoldTrait(tr: Trait): Hint.Native = {
-    Hint.Native(tr.toShapeId, unfoldNode(tr.toNode(), tr.toShapeId()))
+    Hint.Native(
+      tr.toShapeId,
+      cats.Eval.later(unfoldNode(tr.toNode(), tr.toShapeId()))
+    )
   }
 
   private def unfoldNodeAndType(layer: NodeAndType): TypedNode[NodeAndType] =
@@ -1289,8 +1273,8 @@ private[codegen] class SmithyToIR(
       case (N.ObjectNode(map), UnRef(S.Structure(struct))) =>
         val shapeId = struct.getId()
         val ref = Type.Ref(shapeId.getNamespace(), shapeId.getName())
-        val structFields = struct.getFieldsPlain
-        val fieldNames = struct.getFieldsPlain.map(_.name)
+        val structFields = struct.fields
+        val fieldNames = struct.fields.map(_.name)
         val fields: List[TypedNode.FieldTN[NodeAndType]] = structFields.map {
           case Field(_, realName, tpe, mod, _, _)
               if mod.typeMod == Field.TypeModification.None =>