+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Thank you!
 # 0.18.35
 
 * json, documents: Add support for `@jsonUnknown` in unions (Open Unions) in [#1677](https://github.com/disneystreaming/smithy4s/pull/1677)
+* codegen: Prevent `StackOverflowError` in dealing with recursive collection traits in [#1708](https://github.com/disneystreaming/smithy4s/pull/1708)
 
 # 0.18.34